feat: link layer security, SCHC rules 5-6, gateway tunnel auth
Major additions across C/Python/Rust implementations:
- Schnorr48 link signatures and replay protection persistence
- OSCORE context persistence and key update support
- SCHC compression rules 5-6 with MQTT-SN support
- Gateway tunnel authentication and handoff protocol
- Presence/location tracking and confessions resources
- RPL DAO timing, short address assignment, DIS handling
- Extensive cross-implementation test vectors

+++ lichen/subsys/schc/bitstream.c
@@ -54,7 +54,8 @@ int schc_bit_writer_write(struct schc_bit_writer *writer,
 	if (remaining > 0) {
 		uint8_t partial = (uint8_t)(value & ((1U << remaining) - 1));
 
-		writer->buf[writer->nbits / 8] = partial << (8 - remaining);
+		writer->buf[writer->nbits / 8] =
+			(uint8_t)(partial << (8 - remaining));
 		writer->nbits += remaining;
 	}
